Enforce CPU time limits in SqlQueryManager
Previously it was enforced in ClusterMemoryManager. Given that SqlQueryManager
is already enforcing a bunch of other limits and given that this limit is not
related to memory, enforcing it in SqlQueryManager makes more sense.
